fixed delete_user endpoint output

+++ src/main.py
@@ -56,19 +56,30 @@ async def get_users():
 
 
 @app.post("/create_user", response_model=None)
-async def create_user(email: str = Body(...), profile: dict = Body(...), city: str = Body(...)):
+async def create_user(email: str = Body(...),
+                      profile: dict = Body(...),
+                      city: str = Body(...)):
     result = await collection.insert_one(User(email, profile, city).to_dict())
     obj_id = str(result.inserted_id)
     return {"id": obj_id} | User(email, profile, city).to_dict()
 
 
 @app.post("/update_user/{user_id}", response_model=None)
-async def update_user(user_id: str, email: str = Body(...), profile: dict = Body(...), city: str = Body(...)):
+async def update_user(user_id: str,
+                      email: str = Body(...),
+                      profile: dict = Body(...),
+                      city: str = Body(...)):
     await collection.update_one({'_id': ObjectId(user_id)}, {'$set': User(email, profile, city).to_dict()})
     return {"id": user_id} | User(email, profile, city).to_dict()
 
 
 @app.post("/delete_user/{user_id}", response_model=None)
 async def delete_user(user_id: str):
-    await collection.delete_one({'_id': ObjectId(user_id)})
-    return {"user with id - " + str(user_id) + "has been deleted"}
+    target = await collection.find_one({'_id': ObjectId(user_id)})
+    if target is None:
+        return {"User not found"}
+    else:
+        target['_id'] = str(target['_id'])
+        await collection.delete_one({'_id': ObjectId(user_id)})
+
+    return target
